Move beer files to a specific package
