+++ src/main/java/beerduff/api/beer/BeerSetup.java
@@ -1,4 +1,4 @@
-package beerduff.api;
+package beerduff.api.beer;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.context.event.ApplicationReadyEvent;
